+++ frontend/src/components/refine-ui/data-table/data-table-filter.tsx
@@ -17,7 +17,7 @@ import {
   CommandItem,
   CommandList,
 } from "@/components/ui/command";
-import { Input } from "@/components/ui/input";
+import { Input } from "@/components/ui/forms/input";
 import {
   Popover,
   PopoverContent,
